requirements.txt
# FastAPI y dependencias
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
email-validator>=2.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Seguridad
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
PyJWT>=2.8.0

# Base de datos
sqlalchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7
pgvector>=0.2.3

# Redis
redis>=5.0.0
aioredis>=2.0.0

# HTTP y Networking
httpx>=0.24.1
httpcore>=0.18.0

# Procesamiento de Documentos
reportlab>=4.0.4
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
markdown
xhtml2pdf
jinja2

# Utilidades
PyYAML>=6.0.1
Pillow>=10.0.0
numpy>=1.25.2
pandas>=2.1.0
